requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.11.17
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.44.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
openai==1.3.4
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
